Add files via upload
This is a test file to make sure Ollama is working properly and a remote connection is established.  In setting this system up, the user needs to make sure that the firewalls of different computers allow incoming and outgoing connections.  While this gives the URL as "localhost", if you switch it out with an IP address on your local network, it should work.  Also, I have chosen "tinyllama" or "phi3" as my model because my system has minimal computational power.